+++ src/main/java/webserver/WebServer.java
@@ -2,6 +2,7 @@ package webserver;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import webserver.handler.RequestHandler;
 
 import java.net.ServerSocket;
 import java.net.Socket;
@@ -10,8 +11,8 @@ public class WebServer {
     private static final Logger logger = LoggerFactory.getLogger(WebServer.class);
     private static final int DEFAULT_PORT = 8080;
 
-    public static void main(String args[]) throws Exception {
-        int port = 0;
+    public static void main(String[] args) throws Exception {
+        int port;
         if (args == null || args.length == 0) {
             port = DEFAULT_PORT;
         } else {
