+++ src/main/java/webserver/controller/MainController.java
@@ -1,8 +1,9 @@
-package webserver;
+package webserver.controller;
 
 import db.DataBase;
 import model.User;
 import type.ContentType;
+import webserver.parser.Response;
 
 import java.util.Map;
