+++ src/main/java/webserver/handler/RequestHandler.java
@@ -1,7 +1,11 @@
-package webserver;
+package webserver.handler;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import webserver.parser.Request;
+import webserver.parser.Response;
+import webserver.requestmapper.HandlerMapper;
+import webserver.requestmapper.ResourceMapper;
 
 import java.io.*;
 import java.net.Socket;
