+++ src/main/java/webserver/parser/Request.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.parser;
 
 import utils.IOUtils;
