+++ src/main/java/webserver/parser/RequestHeader.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.parser;
 
 import java.util.HashMap;
 import java.util.Map;
