+++ src/main/java/webserver/parser/Response.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.parser;
 
 import type.ContentType;
