+++ src/main/java/webserver/requestmapper/ResourceMapper.java
@@ -1,7 +1,8 @@
-package webserver;
+package webserver.requestmapper;
 
 import type.ContentType;
 import utils.FileIoUtils;
+import webserver.parser.Response;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
